requirements.txt
boto3==1.42.13
botocore==1.42.13
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
greenlet==3.3.0
idna==3.11
jmespath==1.0.1
playwright==1.57.0
proto-plus==1.27.0
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyee==13.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
requests==2.32.5
rsa==4.9.1
s3transfer==0.16.0
six==1.17.0
typing_extensions==4.15.0
urllib3==2.6.2
python-dotenv==1.0.0
msal==1.28.0
